+++ native~/Runtime/src/TextureLoader.h
@@ -17,14 +17,17 @@ namespace CesiumForUnityNative {
 class TextureLoader {
 public:
   static ::DotNet::UnityEngine::Texture
-  loadTexture(const CesiumGltf::ImageCesium& image);
-
-  static ::DotNet::UnityEngine::Texture
-  loadTexture(const CesiumGltf::Model& model, std::int32_t textureIndex);
+  loadTexture(const CesiumGltf::ImageCesium& image, bool sRGB);
 
   static ::DotNet::UnityEngine::Texture loadTexture(
       const CesiumGltf::Model& model,
-      const CesiumGltf::Texture& texture);
+      std::int32_t textureIndex,
+      bool sRGB);
+
+  static ::DotNet::UnityEngine::Texture loadTexture(
+      const CesiumGltf::Model& model,
+      const CesiumGltf::Texture& texture,
+      bool sRGB);
 };
 
 } // namespace CesiumForUnityNative